requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-socketio==5.11.3
python-engineio==4.9.1
twilio==9.3.2
python-dotenv==1.0.1
websockets==12.0
openai==1.43.0
aiohttp>=3.10.2,<4
numpy==2.2.3
flask-sock==0.5.1

